Push finished project
Removed the while loop because it wasn't working the way
I wanted it to and I haven't done the loop lesson yet
anyway. Used else if statements and changed human/computerSelection
variables from const to let so that the values can be reassigned
everytime another round is played. I'm happy with the result ^.^;